backend/requirements.txt
Django>=4.0
psycopg2>=2.8
requests
djangorestframework>=3.14